+++ setup.py
@@ -3,6 +3,6 @@ from setuptools import setup, find_packages
 setup(
     name="ndi",
     version="0.0.1",
-    packages=find_packages(where="ndi"),
+    packages=find_packages(where="code"),
     install_requires=["numpy"]
 )
